+++ otelcollector/shared/configmap/mp/tomlparser-default-scrape-settings.go
@@ -30,7 +30,7 @@ func (fcl *FilesystemConfigLoader) SetDefaultScrapeSettings() (map[string]string
 	return config, nil
 }
 
-func (fcl *FilesystemConfigLoader) ParseConfigMapForDefaultScrapeSettings() (map[string]string, error) {
+func (fcl *FilesystemConfigLoader) ParseConfigMapForDefaultScrapeSettings(metricsConfigBySection map[string]map[string]string, schemaVersion string) (map[string]string, error) {
 	config := make(map[string]string)
 	// Set default values
 	config["kubelet"] = "true"
@@ -51,29 +51,20 @@ func (fcl *FilesystemConfigLoader) ParseConfigMapForDefaultScrapeSettings() (map
 	config["acstor-capacity-provisioner"] = "true"
 	config["acstor-metrics-exporter"] = "true"
 
-	if _, err := os.Stat(fcl.ConfigMapMountPath); os.IsNotExist(err) {
-		fmt.Println("configmap for default scrape settings not mounted, using defaults")
-		return config, nil
+	configSectionName := "default-scrape-settings-enabled"
+	if schemaVersion == "v2" {
+          configSectionName = "default-targets-scrape-enabled"
 	}
-
-	content, err := os.ReadFile(string(fcl.ConfigMapMountPath))
-	if err != nil {
-		return config, fmt.Errorf("using default values, error reading config map file: %s", err)
-	}
-
-	lines := strings.Split(string(content), "\n")
-	for _, line := range lines {
-		parts := strings.SplitN(line, "=", 2)
-		if len(parts) == 2 {
-			key := strings.TrimSpace(parts[0])
-			value := strings.TrimSpace(parts[1])
+	// Override defaults with values from metricsConfigBySection
+	if settings, ok := metricsConfigBySection[configSectionName]; ok {
+		for key, value := range settings {
 			if _, ok := config[key]; ok {
 				config[key] = value
 			}
 		}
 	}
 
-	fmt.Println("using configmap for default scrape settings...")
+	fmt.Println("Using configmap for default scrape settings...")
 	return config, nil
 }
 
@@ -203,7 +194,7 @@ func (fcw *FileConfigWriter) WriteDefaultScrapeSettingsToFile(filename string, c
 	return nil
 }
 
-func (c *Configurator) ConfigureDefaultScrapeSettings() {
+func (c *Configurator) ConfigureDefaultScrapeSettings(metricsConfigBySection map[string]map[string]string) {
 	configSchemaVersion := os.Getenv("AZMON_AGENT_CFG_SCHEMA_VERSION")
 
 	fmt.Printf("Start prometheus-collector-settings Processing\n")
@@ -211,8 +202,8 @@ func (c *Configurator) ConfigureDefaultScrapeSettings() {
 	// Load default settings based on the schema version
 	var defaultSettings map[string]string
 	var err error
-	if configSchemaVersion != "" && strings.TrimSpace(configSchemaVersion) == "v1" {
-		defaultSettings, err = c.ConfigLoader.ParseConfigMapForDefaultScrapeSettings()
+	if configSchemaVersion != "" && (strings.TrimSpace(configSchemaVersion) == "v1" || strings.TrimSpace(configSchemaVersion) == "v2") {
+		defaultSettings, err = c.ConfigLoader.ParseConfigMapForDefaultScrapeSettings(metricsConfigBySection, configSchemaVersion)
 	} else {
 		defaultSettings, err = c.ConfigLoader.SetDefaultScrapeSettings()
 	}
@@ -249,13 +240,20 @@ func (c *Configurator) ConfigureDefaultScrapeSettings() {
 	fmt.Printf("End prometheus-collector-settings Processing\n")
 }
 
-func tomlparserDefaultScrapeSettings() {
+func tomlparserDefaultScrapeSettings(metricsConfigBySection map[string]map[string]string) {
+
+	configSchemaVersion := os.Getenv("AZMON_AGENT_CFG_SCHEMA_VERSION")
+	configLoaderPath := defaultSettingsMountPath
+	if configSchemaVersion != "" && strings.TrimSpace(configSchemaVersion) == "v2" {
+		configLoaderPath = defaultSettingsMountPathv2
+	}
+
 	configurator := &Configurator{
-		ConfigLoader:   &FilesystemConfigLoader{ConfigMapMountPath: defaultSettingsMountPath},
+		ConfigLoader:   &FilesystemConfigLoader{ConfigMapMountPath: configLoaderPath},
 		ConfigWriter:   &FileConfigWriter{},
 		ConfigFilePath: defaultSettingsEnvVarPath,
 		ConfigParser:   &ConfigProcessor{},
 	}
 
-	configurator.ConfigureDefaultScrapeSettings()
+	configurator.ConfigureDefaultScrapeSettings(metricsConfigBySection)
 }