+++ RELEASENOTES.md
@@ -29,6 +29,7 @@
   - fix: set proxy the old way for mdsd in golang (https://github.com/Azure/prometheus-collector/pull/1062)
   - fix: add cluster scope to recording rules in policy (https://github.com/Azure/prometheus-collector/pull/1064)
   - upgrade: Upgrade otelcollector and targetallocator to 0.117.0 (https://github.com/Azure/prometheus-collector/pull/1063)
+  - feat: Configmap update for CCP (v2 + v1 schema support) (https://github.com/Azure/prometheus-collector/pull/1056)
 
 * Pipeline/Docs/Templates Updates:
    * fix: add cluster scope to recording rules in policy (https://github.com/Azure/prometheus-collector/pull/1064)